Add duk_is_heapptr_valid API call.
This is a hack that enables the caller to determine if an object
referenced by a heapptr is on the duktape internal GC finalize list.
If it is, it's invalid, and the heapptr shouldn't be pushed.

This hack steals the ES6 proxy-object heap flag and uses it to mark
whether a heap object is on the finalize list. I had to steal a
flag because all 32 of the flags were already in use, and we don't
use ES6 proxy objects.

This change is temporary, and I'll be able to throw it away once
duktape is modified to support pushing of heapptrs until a finalizer
callback is received. This work is currently underway by the duktape
author.